drm: remove legacy mode_group handling
During initialization of a drm device a struct drm_mode_group is
filled with the encoder/connector/crtc ids from the corresponding
lists, so the drm_mode_group contains the same data as already is
in the lists. Then in drm_mode_getresources either the lists or
the drm_mode_group are used. As both contain the same data, remove
the drm_mode_group and all code around it altogether.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>