rtc: Add SRTC support for FreeScale MX51 and MX53
This is based on the following FreeScale patches:
- ENGR00141217-6 srtc: add rtc driver for 2.6.38
- ENGR00142296-2 SRTC: Upgrade driver to kernel version 2.6.38
- ENGR00141217-2 include file upgrade to 2.6.38
by Frank Li and Anish Trivedi.

Not to clash with rtc-mxc.c if both are compiled in, the driver name was
changed to rtc-imx-v2.c / "imx-srtc", using the imx naming convention
instead of mxc.

Since rtc_device_register calls imx_srtc_read_time, which expects private
driver data to be set, the platform_set_drvdata call was moved before
rtc_device_register.

Signed-off-by: Philipp Zabel <pza@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>